Fix return type of getPercent() method
The Xapian API returns int and the public xapian-glib API does too,
so it doesn't make much sense to convert that to double and back
in this internal layer.

+++ xapian-glib/xapian-mset-iterator.cc
@@ -156,7 +156,7 @@ class IteratorData {
       return mCurrent.get_weight ();
     }
 
-    double getPercent () {
+    int getPercent () {
       if (!mCurrentInitialized)
         return 0;
